requirements.txt
firebase_admin==6.0.1
Flask==2.1.0
Flask_Cors==3.0.10
python-dotenv==0.21.0
web3==5.31.1
Werkzeug==2.0.3
